+++ changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Fixed
+
+- Fixed a wrong text domain in installer which prevents the translation of some texts there
+
 ## [5.7.0] - 07.09.2026
 
 ### Added